chore(refactor): Rename Plural Packages
Rename plural packages to singular
