+++ src/main/kotlin/com/expediagroup/sdk/core/plugin/authentication/strategy/bearer/BearerTokensInfo.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.expediagroup.sdk.core.plugin.authentication.strategies.bearer
+package com.expediagroup.sdk.core.plugin.authentication.strategy.bearer
 
 import com.expediagroup.sdk.core.constant.Authentication
 import com.expediagroup.sdk.core.constant.Constant
