+++ src/test/kotlin/com/expediagroup/sdk/core/plugin/authentication/strategy/signature/SignatureCalculatorTest.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.expediagroup.sdk.core.plugin.authentication.strategies.signature
+package com.expediagroup.sdk.core.plugin.authentication.strategy.signature
 
 import com.expediagroup.sdk.core.commons.TestConstants.CLIENT_KEY_TEST_CREDENTIAL
 import com.expediagroup.sdk.core.commons.TestConstants.CLIENT_SECRET_TEST_CREDENTIAL
